Aggregated metrics tree

+++ code/test-algorithm/src/main/java/tech/dario/dissertation/testalgorithm/Constant.java
@@ -4,21 +4,21 @@ import tech.dario.dissertation.agent.annotations.Measured;
 
 public class Constant {
   @Measured
-  public void quick(int n) {
+  public void quick(long n) {
     for (long i = 0; i < 1562l; i++) {
 
     }
   }
 
   @Measured
-  public void average(int n) {
+  public void average(long n) {
     for (long i = 0; i < 6250l; i++) {
 
     }
   }
 
   @Measured
-  public void slow(int n) {
+  public void slow(long n) {
     for (long i = 0; i < 25000l; i++) {
 
     }
