+++ code/test-algorithm/src/main/java/tech/dario/dissertation/testalgorithm/Executor3.java
@@ -22,7 +22,7 @@ public class Executor3 {
   }
 
   @Measured
-  public void execute(int n) {
+  public void execute(long n) {
     cubic.quick(n);
   }
 }
