+++ code/test-algorithm/src/main/java/tech/dario/dissertation/testalgorithm/Logarithmic.java
@@ -4,21 +4,21 @@ import tech.dario.dissertation.agent.annotations.Measured;
 
 public class Logarithmic {
   @Measured
-  public void quick(int n) {
+  public void quick(long n) {
     for (long i = 0; i < Math.log(n) * 1562l; i++) {
 
     }
   }
 
   @Measured
-  public void average(int n) {
+  public void average(long n) {
     for (long i = 0; i < Math.log(n) * 6250l; i++) {
 
     }
   }
 
   @Measured
-  public void slow(int n) {
+  public void slow(long n) {
     for (long i = 0; i < Math.log(n) * 25000l; i++) {
 
     }
