+++ code/test-algorithm/src/main/java/tech/dario/dissertation/testalgorithm/TestAlgorithm.java
@@ -9,7 +9,7 @@ public class TestAlgorithm {
   private static final Logger LOGGER = LoggerFactory.getLogger(TestAlgorithm.class);
 
   @Measured
-  public void doTask(int n) {
+  public void doTask(long n) {
     LOGGER.info("TestAlgorithm.doTask({})", n);
     new Executor1().execute(n);
   }
