+++ code/time-complexity-analysis-sdk/src/main/java/tech/dario/dissertation/timecomplexityanalysis/sdk/domain/Algorithm.java
@@ -1,5 +1,5 @@
 package tech.dario.dissertation.timecomplexityanalysis.sdk.domain;
 
 public interface Algorithm {
-  void run(int n);
+  void run(long n);
 }
